requirements.txt
Django==2.2.5
django-rest-swagger==2.2.0
djangorestframework==3.10.3
gunicorn==19.9.0
psycopg2-binary==2.8.3
psycopg2==2.8.3
requests==2.22.0